tests: unify fixtures
