+++ tests/tests_integration/test_api/test_hosted_extractors/test_jobs.py
@@ -6,57 +6,41 @@ from cognite.client import CogniteClient
 from cognite.client.data_classes.hosted_extractors import (
     CogniteFormat,
     Destination,
-    DestinationWrite,
-    EventHubSourceWrite,
     Job,
     JobList,
     JobLogsList,
     JobMetricsList,
     JobUpdate,
     JobWrite,
-    SessionWrite,
     Source,
 )
 from cognite.client.exceptions import CogniteAPIError
 from cognite.client.utils._text import random_string
 
-
-@pytest.fixture
-def one_destination(cognite_client: CogniteClient, fresh_session: SessionWrite) -> Destination:
-    my_dest = DestinationWrite(
-        external_id=f"myNewDestination-{random_string(10)}",
-        credentials=fresh_session,
-    )
-    created = cognite_client.hosted_extractors.destinations.create(my_dest)
-    yield created
-
-    cognite_client.hosted_extractors.destinations.delete(created.external_id, ignore_unknown_ids=True)
-
-
-@pytest.fixture
-def one_source(cognite_client: CogniteClient) -> Source:
-    my_source = EventHubSourceWrite(
-        external_id=f"myNewSource-{random_string(10)}",
-        host="myhost",
-        event_hub_name="myeventhub",
-        key_name="mykeyname",
-        key_value="secret",
-    )
-    retrieved = cognite_client.hosted_extractors.sources.retrieve(my_source.external_id, ignore_unknown_ids=True)
-    if retrieved:
-        yield retrieved
-    created = cognite_client.hosted_extractors.sources.create(my_source)
-    yield created
-
-    cognite_client.hosted_extractors.sources.delete(created.external_id, ignore_unknown_ids=True)
+# @pytest.fixture
+# def one_source(cognite_client: CogniteClient) -> Source:
+#     my_source = EventHubSourceWrite(
+#         external_id=f"myNewSource-{random_string(10)}",
+#         host="myhost",
+#         event_hub_name="myeventhub",
+#         key_name="mykeyname",
+#         key_value="secret",
+#     )
+#     retrieved = cognite_client.hosted_extractors.sources.retrieve(my_source.external_id, ignore_unknown_ids=True)
+#     if retrieved:
+#         yield retrieved
+#     created = cognite_client.hosted_extractors.sources.create(my_source)
+#     yield created
+#
+#     cognite_client.hosted_extractors.sources.delete(created.external_id, ignore_unknown_ids=True)
 
 
 @pytest.fixture
-def one_job(cognite_client: CogniteClient, one_source: Source, one_destination: Destination) -> Job:
+def one_job(cognite_client: CogniteClient, one_event_hub_source: Source, one_destination: Destination) -> Job:
     my_job = JobWrite(
         external_id=f"myJobForTesting-{random_string(10)}",
         destination_id=one_destination.external_id,
-        source_id=one_source.external_id,
+        source_id=one_event_hub_source.external_id,
         format=CogniteFormat(encoding="utf16"),
     )
     retrieved = cognite_client.hosted_extractors.jobs.retrieve(my_job.external_id, ignore_unknown_ids=True)
@@ -70,12 +54,12 @@ def one_job(cognite_client: CogniteClient, one_source: Source, one_destination: 
 
 class TestJobs:
     def test_create_update_retrieve_delete(
-        self, cognite_client: CogniteClient, one_destination: Destination, one_source: Source
+        self, cognite_client: CogniteClient, one_destination: Destination, one_event_hub_source: Source
     ) -> None:
         my_job = JobWrite(
             external_id=f"myJobForTesting-{random_string(10)}",
             destination_id=one_destination.external_id,
-            source_id=one_source.external_id,
+            source_id=one_event_hub_source.external_id,
             format=CogniteFormat(
                 encoding="utf16",
             ),
@@ -111,12 +95,12 @@ class TestJobs:
         assert isinstance(res, JobList)
 
     def test_update_using_write_object(
-        self, cognite_client: CogniteClient, one_destination: Destination, one_source: Source
+        self, cognite_client: CogniteClient, one_destination: Destination, one_event_hub_source: Source
     ) -> None:
         my_job = JobWrite(
             external_id=f"myJobForTesting-{random_string(10)}",
             destination_id=one_destination.external_id,
-            source_id=one_source.external_id,
+            source_id=one_event_hub_source.external_id,
             format=CogniteFormat(
                 encoding="utf16",
             ),
@@ -128,7 +112,7 @@ class TestJobs:
             update = JobWrite(
                 external_id=my_job.external_id,
                 destination_id=one_destination.external_id,
-                source_id=one_source.external_id,
+                source_id=one_event_hub_source.external_id,
                 format=CogniteFormat(
                     encoding="utf16le",
                 ),
