+++ tests/tests_integration/test_api/test_hosted_extractors/test_sources.py
@@ -1,7 +1,5 @@
 from __future__ import annotations
 
-import platform
-
 import pytest
 
 from cognite.client import CogniteClient
@@ -13,21 +11,6 @@ from cognite.client.data_classes.hosted_extractors import (
 )
 from cognite.client.exceptions import CogniteAPIError
 from cognite.client.utils._text import random_string
-
-
-@pytest.fixture(scope="session")
-def one_event_hub_source(cognite_client: CogniteClient) -> SourceList:
-    my_hub = EventHubSourceWrite(
-        external_id=f"myNewHub-{platform.system()}-{platform.python_version()}",
-        host="myHost",
-        key_name="myKeyName",
-        key_value="myKey",
-        event_hub_name="myEventHub",
-    )
-    retrieved = cognite_client.hosted_extractors.sources.retrieve([my_hub.external_id], ignore_unknown_ids=True)
-    if retrieved:
-        return retrieved
-    return cognite_client.hosted_extractors.sources.create([my_hub])
 
 
 class TestSources:
